# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tracking Research Status API
  version: 1.0.0
  description: |
    API contract for the Tracking Research Status MVP.

    The API has no authentication or role-specific views. Every caller sees the
    same research list. UI pages for viewing or changing status and process are
    outside MVP scope, but the status and process API operations are included.

    Every error response uses `{"error":{"code":"...","message":"..."}}` and
    does not expose stack traces or database details. Unknown routes return
    `404 ROUTE_NOT_FOUND`; unsupported methods on known routes return
    `405 METHOD_NOT_ALLOWED`.

    POST and PUT accept multipart/form-data with one PDF contract file; PATCH
    accepts one JSON object. Media types are matched case-insensitively. JSON
    duplicate keys, unknown fields, and non-object top-level values are
    rejected. The PDF limit is 20 MiB and the PATCH JSON limit is 64 KiB.
security: []
servers:
  - url: /
...
paths:
...
  /api/v1/researches:
    get:
      tags:
        - Researches
      operationId: listResearches
      summary: List all researches
      description: |
        Returns the shared research list sorted by title ascending and then by
        ID ascending when titles are equal. An empty database returns `[]`.
        This operation accepts neither a non-empty request body nor query
        parameters.
      responses:
        "200":
          description: The complete research list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Research"
        "400":
          $ref: "#/components/responses/InvalidRequestBody"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags:
        - Researches
      operationId: createResearch
      summary: Create a research
      description: |
        Creates one research atomically from multipart form data. The server generates a positive,
        immutable, unique ID that is never reused after deletion. Status starts
        at "กำลังดำเนินการ" and process starts at "สัญญาโครงการ".

        continuationOfId is required. A root research sends null; a continuation
        sends the positive ID of an existing research. The referenced research
        may already be completed or terminated. A root title cannot duplicate
        any existing trimmed title. Continuations may use duplicate titles.
        Title comparison is case-sensitive after Unicode trimming. Every required
        form part occurs exactly once; unknown or duplicate parts, including a
        projectMembers part that is not one JSON array, are validation errors.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateResearchRequest"
            encoding:
              projectMembers:
                contentType: application/json
              contractFile:
                contentType: application/pdf
      responses:
        "201":
          description: The research was created and persisted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Research"
        "404":
          $ref: "#/components/responses/ContinuationNotFound"
        "409":
          $ref: "#/components/responses/TitleAlreadyExists"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/researches/{id}:
    parameters:
      - $ref: "#/components/parameters/ResearchId"

    put:
      tags:
        - Researches
      operationId: updateResearch
      summary: Replace research details
      description: |
        Atomically replaces all editable data and the contract PDF. ID,
        continuationOfId, researchKind, status, and process remain unchanged. A root may retain its current title
        while changing other data even when a continuation has the same title.
        If a root changes its title, the new trimmed title cannot duplicate any
        existing research title. A continuation may change to a duplicate title.
        Every required form part occurs exactly once; unknown or duplicate parts,
        including a projectMembers part that is not one JSON array, are validation
        errors.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateResearchRequest"
            encoding:
              projectMembers:
                contentType: application/json
              contractFile:
                contentType: application/pdf
      responses:
        "200":
          description: The research details were replaced successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Research"
        "404":
          $ref: "#/components/responses/ResearchNotFound"
        "409":
          $ref: "#/components/responses/TitleAlreadyExists"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags:
        - Researches
      operationId: deleteResearch
      summary: Delete a research
      description: |
        Deletes the research identified by ID. The deleted ID is never reused.
        A research referenced through continuationOfId cannot be deleted. This
        operation accepts neither a non-empty request body nor query parameters.
      responses:
        "204":
          description: The research was deleted successfully; the response has no body.
        "400":
          $ref: "#/components/responses/InvalidRequestBody"
        "404":
          $ref: "#/components/responses/ResearchNotFound"
        "409":
          $ref: "#/components/responses/ResearchHasContinuations"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/researches/{id}/status:
    parameters:
      - $ref: "#/components/parameters/ResearchId"

    patch:
      tags:
        - Researches
      operationId: updateResearchStatus
      summary: Advance research status
      description: |
        Updates status atomically from the latest persisted value. A normal
        status may move only to the next normal status, or jump directly to
        "โครงการเสร็จสิ้น" or "ยุติโครงการ". Normal statuses cannot be skipped
        or reversed. Sending the current status again is idempotent and returns
        200, including the same terminal status.

        Both terminal statuses end the project. Entering a terminal status keeps
        the current process unchanged and does not mark it complete. After that,
        status cannot change to another value and process cannot advance.
      requestBody:
        required: true
        x-max-body-bytes: 65536
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status was advanced or the idempotent value was accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Research"
        "400":
          $ref: "#/components/responses/InvalidJson"
        "404":
          $ref: "#/components/responses/ResearchNotFound"
        "409":
          $ref: "#/components/responses/StatusConflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/researches/{id}/process:
    parameters:
      - $ref: "#/components/parameters/ResearchId"

    patch:
      tags:
        - Researches
      operationId: updateResearchProcess
      summary: Advance research process
      description: |
        Updates the single current process atomically from the latest persisted
        value. Process may move only to the next value; it cannot skip, reverse,
        or advance beyond "การปิดบัญชีธนาคาร". Sending the current process again
        is idempotent and returns 200.

        Advancing process does not change status. Completing the final process
        does not automatically complete the project. Process cannot be changed
        after status becomes "โครงการเสร็จสิ้น" or "ยุติโครงการ".
      requestBody:
        required: true
        x-max-body-bytes: 65536
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProcessRequest"
      responses:
        "200":
          description: Process was advanced or the idempotent value was accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Research"
        "400":
          $ref: "#/components/responses/InvalidJson"
        "404":
          $ref: "#/components/responses/ResearchNotFound"
        "409":
          $ref: "#/components/responses/ProcessConflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    ResearchId:
      name: id
      in: path
      required: true
      description: Positive, immutable server-generated research ID.
      schema:
        $ref: "#/components/schemas/ResearchId"
  schemas:
...
    Research:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - continuationOfId
        - isSubsidized
        - projectMembers
        - fundingType
        - fundingSourceName
        - contractNumber
        - contractFile
        - projectType
        - researchKind
        - responsibleProjectUnit
        - responsibleBudgetUnit
        - startDate
        - endDate
        - budgetAmount
        - thaiAbstract
        - englishAbstract
        - objectives
        - keywords
        - status
        - process
      properties:
        id:
          $ref: "#/components/schemas/ResearchId"
        title:
          $ref: "#/components/schemas/ResearchTitle"
        continuationOfId:
          $ref: "#/components/schemas/NullableResearchId"
        isSubsidized:
          type: boolean
        projectMembers:
          type: array
          items:
            $ref: "#/components/schemas/ResearchMember"
        fundingType:
          $ref: "#/components/schemas/FundingType"
        fundingSourceName:
          $ref: "#/components/schemas/ResearchText"
        contractNumber:
          $ref: "#/components/schemas/ResearchText"
        contractFile:
          $ref: "#/components/schemas/ContractFileMetadata"
        projectType:
          $ref: "#/components/schemas/ProjectType"
        researchKind:
          $ref: "#/components/schemas/ResearchKind"
        responsibleProjectUnit:
          $ref: "#/components/schemas/ResearchText"
        responsibleBudgetUnit:
          $ref: "#/components/schemas/ResearchText"
        startDate:
          $ref: "#/components/schemas/BuddhistDate"
        endDate:
          $ref: "#/components/schemas/BuddhistDate"
        budgetAmount:
          $ref: "#/components/schemas/BudgetAmount"
        thaiAbstract:
          $ref: "#/components/schemas/ResearchText"
        englishAbstract:
          $ref: "#/components/schemas/ResearchText"
        objectives:
          $ref: "#/components/schemas/ResearchText"
        keywords:
          $ref: "#/components/schemas/ResearchText"
        status:
          $ref: "#/components/schemas/ResearchStatus"
        process:
          $ref: "#/components/schemas/ResearchProcess"
    CreateResearchRequest:
      type: object
      additionalProperties: false
      required:
        - title
        - continuationOfId
        - isSubsidized
        - projectMembers
        - fundingType
        - fundingSourceName
        - contractNumber
        - contractFile
        - projectType
        - researchKind
        - responsibleProjectUnit
        - responsibleBudgetUnit
        - startDate
        - endDate
        - budgetAmount
        - thaiAbstract
        - englishAbstract
        - objectives
        - keywords
      properties:
        title:
          $ref: "#/components/schemas/ResearchTitle"
        continuationOfId:
          $ref: "#/components/schemas/NullableResearchId"
        isSubsidized:
          type: boolean
        projectMembers:
          type: array
          minItems: 2
          description: |
            Exactly one form part containing a JSON array. The array must contain
            at least one LEAD and at least one CO_RESEARCHER.
          items:
            $ref: "#/components/schemas/ResearchMember"
        fundingType:
          $ref: "#/components/schemas/FundingType"
        fundingSourceName:
          $ref: "#/components/schemas/ResearchText"
        contractNumber:
          $ref: "#/components/schemas/ResearchText"
        contractFile:
          $ref: "#/components/schemas/ContractFile"
        projectType:
          $ref: "#/components/schemas/ProjectType"
        researchKind:
          $ref: "#/components/schemas/ResearchKind"
        responsibleProjectUnit:
          $ref: "#/components/schemas/ResearchText"
        responsibleBudgetUnit:
          $ref: "#/components/schemas/ResearchText"
        startDate:
          $ref: "#/components/schemas/BuddhistDate"
        endDate:
          $ref: "#/components/schemas/BuddhistDate"
        budgetAmount:
          $ref: "#/components/schemas/BudgetAmount"
        thaiAbstract:
          $ref: "#/components/schemas/ResearchText"
        englishAbstract:
          $ref: "#/components/schemas/ResearchText"
        objectives:
          $ref: "#/components/schemas/ResearchText"
        keywords:
          $ref: "#/components/schemas/ResearchText"
    UpdateResearchRequest:
      type: object
      additionalProperties: false
      required:
        - title
        - isSubsidized
        - projectMembers
        - fundingType
        - fundingSourceName
        - contractNumber
        - contractFile
        - projectType
        - responsibleProjectUnit
        - responsibleBudgetUnit
        - startDate
        - endDate
        - budgetAmount
        - thaiAbstract
        - englishAbstract
        - objectives
        - keywords
      properties:
        title:
          $ref: "#/components/schemas/ResearchTitle"
        isSubsidized:
          type: boolean
        projectMembers:
          type: array
          minItems: 2
          description: |
            Exactly one form part containing a JSON array. The array must contain
            at least one LEAD and at least one CO_RESEARCHER.
          items:
            $ref: "#/components/schemas/ResearchMember"
        fundingType:
          $ref: "#/components/schemas/FundingType"
        fundingSourceName:
          $ref: "#/components/schemas/ResearchText"
        contractNumber:
          $ref: "#/components/schemas/ResearchText"
        contractFile:
          $ref: "#/components/schemas/ContractFile"
        projectType:
          $ref: "#/components/schemas/ProjectType"
        responsibleProjectUnit:
          $ref: "#/components/schemas/ResearchText"
        responsibleBudgetUnit:
          $ref: "#/components/schemas/ResearchText"
        startDate:
          $ref: "#/components/schemas/BuddhistDate"
        endDate:
          $ref: "#/components/schemas/BuddhistDate"
        budgetAmount:
          $ref: "#/components/schemas/BudgetAmount"
        thaiAbstract:
          $ref: "#/components/schemas/ResearchText"
        englishAbstract:
          $ref: "#/components/schemas/ResearchText"
        objectives:
          $ref: "#/components/schemas/ResearchText"
        keywords:
          $ref: "#/components/schemas/ResearchText"
    UpdateStatusRequest:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/ResearchStatus"
...
  responses:
    ContinuationNotFound:
      description: continuationOfId does not reference an existing research.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: CONTINUATION_NOT_FOUND
              message: Continuation research was not found.
    InternalError:
      description: An unexpected database or internal error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: An unexpected internal error occurred.
    InvalidJson:
      description: |
        The body is empty or whitespace-only, malformed, has trailing data, or
        contains more than one JSON value.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INVALID_JSON
              message: Request body must contain one valid JSON object.
    InvalidRequestBody:
      description: This operation does not accept a non-empty request body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INVALID_REQUEST_BODY
              message: This operation does not accept a request body.
    PayloadTooLarge:
      description: |
        The request payload exceeds the operation limit: 20 MiB for a contract
        PDF in POST/PUT, or 64 KiB for a PATCH JSON body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: PAYLOAD_TOO_LARGE
              message: Request payload exceeds the allowed size.
...
    ResearchHasContinuations:
      description: The research is still referenced by one or more continuations.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RESEARCH_HAS_CONTINUATIONS
              message: Research has continuations and cannot be deleted.
    ResearchNotFound:
      description: No research has the supplied positive ID.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RESEARCH_NOT_FOUND
              message: Research was not found.
...
    StatusConflict:
      description: |
        The requested status transition skips or reverses a normal status, or
        attempts to change a project that has already ended.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidTransition:
              value:
                error:
                  code: INVALID_STATUS_TRANSITION
                  message: The requested status transition is invalid.
            projectEnded:
              value:
                error:
                  code: PROJECT_ALREADY_ENDED
                  message: The project has already ended.
    TitleAlreadyExists:
      description: A root research would duplicate an existing case-sensitive trimmed title.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: TITLE_ALREADY_EXISTS
              message: A research with this title already exists.
    UnsupportedMediaType:
      description: Content-Type is missing or does not match the operation's required media type.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNSUPPORTED_MEDIA_TYPE
              message: Content-Type is not supported for this operation.
    ValidationError:
      description: |
        The path ID, request object, fields, duplicate keys, unknown fields, or
        query parameters violate this contract.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed.